requirements.txt
asyncpg==0.28.0
setuptools==68.0.0
pydantic==2.1.1
pydantic-settings==2.0.2
fastapi==0.100.0
SQLAlchemy==2.0.19
gunicorn==21.2.0
uvicorn==0.23.1
click==8.1.6
h11==0.14.0
loguru==0.7.0
alembic==1.11.1
psycopg2-binary==2.9.6
python-multipart==0.0.6
pillow==10.0.0
aioboto3==11.2.0
pyjwt
passlib
